change type of the end token

+++ demos/text_detection_demo/cpp/include/cnn.hpp
@@ -46,7 +46,7 @@ class EncoderDecoderCNN : public Cnn {
                       const std::string in_dec_symbol_name,
                       const std::string out_dec_symbol_name,
                       const std::string logits_name,
-                      unsigned end_token,
+                      size_t end_token,
                       const cv::Size &new_input_resolution = cv::Size()
                       );
     InferenceEngine::BlobMap Infer(const cv::Mat &frame) override;
